backend/ML_Deployment/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
keras==2.9.0
numpy==1.23.0
pandas==1.4.3
sklearn==0.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0
scikit-learn>=1.0.0
scikeras==0.8.0
django-environ==0.9.0
tokenizers==0.12.1
xgboost==1.6.1
bs4==0.0.1
python-whois==0.8.0
lxml==4.9.1
tqdm==4.64.0
djangorestframework-api-key==2.2.0
djongo==1.3.6
pymongo==3.12.1
dnspython==2.2.1
sqlparse==0.2.4
pytesseract==0.3.9
ISR==2.2.0
pdf2image==1.16.0
opencv-python==4.6.0.66
regex==2022.7.25
Pillow==9.2.0
img2pdf==0.4.4
